change float -> double of column total_order

+++ database/migrations/2023_10_05_092504_create_orders_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->text('message')->nullable();
             $table->string('code')->nullable();
             $table->tinyInteger('status')->default(1);
-            $table->float('total_order')->nullable();
+            $table->double('total_order', 15, 0)->nullable();
             $table->timestamps();
         });
     }
